Some nice API improvements to render pkg
